+++ app/components/admin/landing-page/AdminTop.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import ViewAddBuildings from './ViewAddBuildings'
+
+export default class AdminTop extends React.Component {
+
+  render() {
+    return (
+      <div className='admin-top'>
+        <ViewAddBuildings />
+      </div>
+    )
+  }
+}
