+++ app/components/admin/landing-page/LandingPage.jsx
@@ -1,81 +1,24 @@
 import React from 'react'
-import Tabs from './tabs/Tabs'
-import Pages from './Pages'
-import Users from './Users'
+import SuperadminTop from './SuperadminTop'
+import AdminTop from './AdminTop'
 import Filters from '../../Filters'
 import Search from '../../Search'
 import api from '../../../../config'
-import { Link } from 'react-router'
-
-const pages = [
-  {
-    label: 'About',
-    route: '/admin/about'
-  },
-  {
-    label: 'Contact',
-    route: '/admin/contact'
-  },
-  {
-    label: 'Glossary',
-    route: '/admin/glossary'
-  },
-]
 
 export default class LandingPage extends React.Component {
-  constructor(props) {
-    super(props)
 
-    this.state = {
-      tab: 'users',  // {'users' | 'pages'}
-      users: []
-    }
-
-    this.changeTab = this.changeTab.bind(this)
-    this.processUsers = this.processUsers.bind(this)
-  }
-
-  changeTab(tab) {
-    this.setState({tab: tab})
-  }
-
-  componentDidMount() {
-    api.get('users', this.processUsers)
-  }
-
-  processUsers(err, res) {
-    if (err) console.warn(err)
-    this.setState({users: res.body})
-  }
-
+  // Protect resources to which only superadmins have access
   render() {
-    const view = this.state.tab == 'users' ?
-        <Users users={this.state.users} />
-      : <Pages pages={pages} />
+    let top = null;
+    api.get('session', (err, res) => {
+      if (err) console.warn(err)
+      top = res.body.session.superadmin ? <SuperadminTop /> : <AdminTop />
+    })
 
     return (
       <div className='admin-landing-page'>
         <div className='top'>
-          <div className='left'>
-            <Tabs tab={this.state.tab} changeTab={this.changeTab} />
-            {view}
-          </div>
-
-          <div className='right'>
-            <div className='darker-shade' />
-            <div className='right-container'>
-              <h1>View Buildings</h1>
-              <div className='admin-search-container'>
-                <div className='admin-search-content'>
-                  <input className='admin-search building-search' placeholder='Search'></input>
-                  <div className='admin-search-button'/>
-                </div>
-              </div>
-              <Link to='/admin/form'>
-                <div className='white-button add-a-building'>Add a Building</div>
-              </Link>
-            </div>
-          </div>
+          {top}
         </div>
         <div className='bottom'>
           <Search />
