+++ app/styles/admin-landing-page.css
@@ -132,92 +132,6 @@
 }
 
 /**
-* Top Right
-**/
-
-.admin-landing-page .top .right {
-  width: 50%;
-  height: 100%;
-  display: inline-block;
-  background: url(/assets/images/new-haven-overhead.jpg);
-  background-repeat: no-repeat;
-  background-position: 50% 10%; 
-  -webkit-background-size: cover;
-  -moz-background-size: cover;
-  -o-background-size: cover;
-  background-size: cover;
-  position: absolute;
-  top: 0;
-  right: 0;
-}
-
-.admin-landing-page .top .right .darker-shade {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background: rgba(0,0,0,0.3);
-}
-
-.admin-landing-page .right-container {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  text-align: center;
-}
-
-.admin-landing-page .right-container h1 {
-  color: #fff;
-  margin-top: 40px;
-  font-family: Tinos;
-}
-
-.admin-landing-page .admin-search-container {
-  text-align: center;
-  width: 100%;
-  margin: 50px 0 30px;
-}
-
-.admin-landing-page .admin-search-content {
-  position: relative;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  overflow: hidden;
-  border-radius: 5px;
-  width: 400px;
-  display: inline-block;
-}
-
-.admin-landing-page .admin-search {
-  width: 100%;
-  padding: 10px;
-  padding-right: 78px;
-  height: 40px;
-  border: none;
-  font-size: 1em;
-  box-sizing: border-box;
-}
-
-.admin-landing-page .admin-search-button {
-  background: #EDA138;
-  width: 70px;
-  height: 100%;
-  position: absolute;
-  top: 0;
-  right: 0;
-  background-image: url(/assets/images/search-icon-white.png);
-  background-repeat: no-repeat;
-  background-position: center 6px;
-  background-size: 27px;
-  cursor: pointer;
-}
-
-/**
 * Filters
 **/
