+++ app/styles/styles.js
@@ -25,3 +25,4 @@ import './related.css'
 import './search.css'
 import './simple-pages.css'
 import './suggest-edit.css'
+import './view-add-buildings.css'
