+++ app/styles/view-add-buildings.css
@@ -1,0 +1,93 @@
+/**
+* AdminLandingPage: View/Add Buildings
+**/
+
+.admin-landing-page .right-container {
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  text-align: center;
+}
+
+.view-add-buildings {
+  height: 100%;
+  display: inline-block;
+  background: url(/assets/images/new-haven-overhead.jpg);
+  background-repeat: no-repeat;
+  background-position: 50% 10%; 
+  -webkit-background-size: cover;
+  -moz-background-size: cover;
+  -o-background-size: cover;
+  background-size: cover;
+  position: absolute;
+  top: 0;
+  right: 0;
+}
+
+/* conditionalize element width on user's status */
+.admin-top .view-add-buildings {
+  width: 100%;
+}
+
+.superadmin-top .view-add-buildings {
+  width: 50%;
+}
+
+.view-add-buildings .darker-shade {
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  background: rgba(0,0,0,0.3);
+}
+
+.view-add-buildings h1 {
+  color: #fff;
+  margin-top: 40px;
+  font-family: Tinos;
+}
+
+.view-add-buildings .admin-search-container {
+  text-align: center;
+  width: 100%;
+  margin: 50px 0 30px;
+}
+
+.view-add-buildings .admin-search-content {
+  position: relative;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  overflow: hidden;
+  border-radius: 5px;
+  width: 400px;
+  display: inline-block;
+}
+
+.view-add-buildings .admin-search {
+  width: 100%;
+  padding: 10px;
+  padding-right: 78px;
+  height: 40px;
+  border: none;
+  font-size: 1em;
+  box-sizing: border-box;
+}
+
+.view-add-buildings .admin-search-button {
+  background: #EDA138;
+  width: 70px;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  right: 0;
+  background-image: url(/assets/images/search-icon-white.png);
+  background-repeat: no-repeat;
+  background-position: center 6px;
+  background-size: 27px;
+  cursor: pointer;
+}
